+++ widgets/src/divider.rs
@@ -2,7 +2,7 @@ use ribir_core::prelude::*;
 
 use crate::prelude::*;
 
-/// Divider is a thin horizontal or vertical line, with padding on either side.
+/// Divider is a thin horizontal or vertical line, with indent on either side.
 ///
 /// # example
 /// ```
@@ -13,89 +13,83 @@ use crate::prelude::*;
 /// let widget = fn_widget! {
 ///   @Column {
 ///     @SizedBox { size: Size::new(10., 0.) }
-///     @Divider { extent: 20. }
+///     @Divider { }
 ///     @SizedBox { size: Size::new(10., 0.) }
 ///   }
 /// };
 ///
-/// // use custom settings
+/// // with indent settings
 /// let widget = fn_widget! {
 ///   @Column {
 ///     @SizedBox { size: Size::new(10., 0.) }
 ///     @Divider {
-///       extent: 20.,
-///       color: Color::RED,
-///       direction: Direction::Horizontal,
-///       // Thickness of line
-///       thickness: 2.,
-///       // front indentation distance
-///       indent: 10.,
-///       // behind indentation distance
-///       end_indent: 10.,
+///       indent: DividerIndent::Start,
 ///     }
 ///     @SizedBox { size: Size::new(10., 0.) }
 ///   }
 /// };
 /// ```
+
+#[derive(Default, Clone, Copy)]
+pub enum DividerIndent {
+  /// Does not indent the divider.
+  #[default]
+  None,
+  /// Indents the divider with equal indent on the leading sides.
+  Start,
+  /// Indents the divider with equal indent on the trailing sides.
+  End,
+  /// Indents the divider with equal indent on the both sides.
+  Both,
+}
+
+class_names! {
+  #[doc = "class name for horizontal divider"]
+  HORIZONTAL_DIVIDER,
+  #[doc = "class name for horizontal divider with inset at leading"]
+  HORIZONTAL_DIVIDER_INDENT_START,
+  #[doc = "class name for horizontal divider with inset at trailing"]
+  HORIZONTAL_DIVIDER_INDENT_END,
+  #[doc = "class name for horizontal divider with inset at both sides"]
+  HORIZONTAL_DIVIDER_INDENT_BOTH,
+  #[doc = "class name for vertical divider"]
+  VERTICAL_DIVIDER,
+  #[doc = "class name for vertical divider with inset at leading"]
+  VERTICAL_DIVIDER_INDENT_START,
+  #[doc = "class name for vertical divider with inset at trailing"]
+  VERTICAL_DIVIDER_INDENT_END,
+  #[doc = "class name for vertical divider with inset at both sides"]
+  VERTICAL_DIVIDER_INDENT_BOTH,
+}
+
 #[derive(Default, Declare)]
 pub struct Divider {
-  #[declare(default = 1.)]
-  // Extent of divider
-  pub extent: f32,
-  // Color of divider
-  #[declare(default=Palette::of(BuildCtx::get()).outline_variant())]
-  pub color: Brush,
-  // Direction of divider
-  #[declare(default=Direction::Horizontal)]
-  pub direction: Direction,
-  // Thickness of line
-  #[declare(default = 1.)]
-  pub thickness: f32,
-  // front indentation distance
-  #[declare(default = 0.)]
-  pub indent: f32,
-  // behind indentation distance
-  #[declare(default = 0.)]
-  pub end_indent: f32,
+  #[declare(default)]
+  indent: DividerIndent,
+
+  #[declare(default = Direction::Horizontal)]
+  direction: Direction,
 }
 
-impl Render for Divider {
-  fn perform_layout(&self, clamp: BoxClamp, _: &mut LayoutCtx) -> Size {
-    if self.direction.is_horizontal() {
-      let width = clamp.max.width;
-      if width.is_finite() { Size::new(width, self.extent) } else { clamp.min }
-    } else {
-      let height = clamp.max.height;
-      if height.is_finite() { Size::new(self.extent, height) } else { clamp.min }
+impl Compose for Divider {
+  fn compose(this: impl StateWriter<Value = Self>) -> Widget<'static> {
+    fn_widget! {
+      @Void {
+        class: pipe!(($this.indent, $this.direction)).map(|(s,d)| {
+          match (s, d) {
+            (DividerIndent::None, Direction::Horizontal) => HORIZONTAL_DIVIDER,
+            (DividerIndent::Start, Direction::Horizontal) => HORIZONTAL_DIVIDER_INDENT_START,
+            (DividerIndent::End, Direction::Horizontal) => HORIZONTAL_DIVIDER_INDENT_END,
+            (DividerIndent::Both, Direction::Horizontal) => HORIZONTAL_DIVIDER_INDENT_BOTH,
+            (DividerIndent::None, Direction::Vertical) => VERTICAL_DIVIDER,
+            (DividerIndent::Start, Direction::Vertical) => VERTICAL_DIVIDER_INDENT_START,
+            (DividerIndent::End, Direction::Vertical) => VERTICAL_DIVIDER_INDENT_END,
+            (DividerIndent::Both, Direction::Vertical) => VERTICAL_DIVIDER_INDENT_BOTH,
+          }
+        }),
+
+      }
     }
-  }
-
-  fn paint(&self, ctx: &mut PaintingCtx) {
-    let rect = self.paint_rect(ctx.box_size().unwrap());
-    let painter = ctx.painter();
-    painter.set_fill_brush(self.color.clone());
-    painter.rect(&rect);
-    painter.fill();
-  }
-
-  fn visual_box(&self, ctx: &mut VisualCtx) -> Option<Rect> {
-    let rect = self.paint_rect(ctx.box_size()?);
-    Some(rect)
-  }
-}
-
-impl Divider {
-  fn paint_rect(&self, mut box_size: Size) -> Rect {
-    if self.direction.is_horizontal() {
-      box_size.width -= self.indent + self.end_indent;
-      box_size.height = self.thickness;
-      let y = (self.extent - self.thickness) / 2.;
-      Rect::new(Point::new(self.indent, y), box_size)
-    } else {
-      box_size.width = self.thickness;
-      box_size.height -= self.indent + self.end_indent;
-      let x = (self.extent - self.thickness) / 2.;
-      Rect::new(Point::new(x, self.indent), box_size)
-    }
+    .into_widget()
   }
 }